extract type resolution helpers from emitter
